+++ chp10_nn/xor/code/src/nn/InputNeuron.java
@@ -16,7 +16,7 @@ public class InputNeuron extends Neuron {
         super(i);
     }
 
-    public void input(float d) {
+    public void input(double d) {
         output = d;
     }
